Convert instrument ID type to i32
instruments.id and instruments.market_id are INTEGER (int4) in Postgres.
Switch Instrument.id/market_id, the InstrumentStore id parameters, and the
market/instrument id plumbing through the trade + resolve flows (button and
modal id parsing, calc helpers, FullMarket::get_instrument) to i32, dropping
the compensating CAST(... AS BIGINT) casts. Calls into the still-i64 order
and position stores get a temporary .into(); the two spots where an i64
Position.instrument_id meets instrument code get a temporary `as i32`.

+++ bot/src/command/leaderboard.rs
@@ -111,7 +111,7 @@ pub async fn run(
         .instrument_store
         .get_all_open_instruments_with_share_counts(&mut tx)
         .await?;
-    let mut instruments_by_market: HashMap<i64, Vec<InstrumentWithShares>> = HashMap::new();
+    let mut instruments_by_market: HashMap<i32, Vec<InstrumentWithShares>> = HashMap::new();
     for i in instruments {
         instruments_by_market
             .entry(i.0.market_id)
